% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{makeGroups}
\alias{makeGroups}
\title{Make a matrix suitable for use with RUVSeq methods such as \code{\link{RUVs}}.}
\usage{
makeGroups(xs)
}
\arguments{
\item{xs}{A vector indicating membership in a group.}
}
\description{
Each row in the returned matrix corresponds to a set of replicate samples.
The number of columns is the size of the largest set of replicates; rows for
smaller sets are padded with -1 values.
}
\examples{
makeGroups(c("A","B","B","C","C","D","D","D","A"))
}
\author{
Kamil Slowikowski
}
\seealso{
\code{\link{RUVs}}
}

